revert: default back to bf16 — fp16 causes 7x decode slowdown
fp16 compute causes MLX to recompile quantized_matmul kernels for
fp16 inputs, resulting in ~105ms/step instead of ~15ms. This is
not a transport issue — the model forward itself is 7x slower.
MLX's quantized kernels are optimized for bf16 inputs only.

The fp16 infrastructure remains for future use if MLX adds fp16
quantized kernel support.

+++ src/exo/worker/engines/mlx/patches/qwen3_5_moe/common.py
@@ -29,7 +29,7 @@ from mlx_lm.models.qwen3_next import Qwen3NextSparseMoeBlock
 # EXO_COMPUTE_DTYPE controls the activation/accumulation dtype for inference.
 # "fp16" (default): ~7% faster quantized_matmul on M4, same gather_qmm perf.
 # "bf16": original dtype matching safetensors storage.
-_DTYPE_STR = os.environ.get("EXO_COMPUTE_DTYPE", "fp16")
+_DTYPE_STR = os.environ.get("EXO_COMPUTE_DTYPE", "bf16")
 COMPUTE_DTYPE: mx.Dtype = mx.bfloat16 if _DTYPE_STR == "bf16" else mx.float16
 METAL_HALF_TYPE: str = "bfloat16_t" if _DTYPE_STR == "bf16" else "half"
